Move IBC crate into component crate
Fix IBC test by using `include_str!` instead of `fs::read_to_string`
