+++ Cargo.toml
@@ -9,7 +9,6 @@ members = [
   "component",
   "crypto",
   "custody",
-  "ibc",
   "storage",
   "tct",
   "tct-property-test",
