+++ component/Cargo.toml
@@ -35,10 +35,20 @@ bech32 = "0.8"
 regex = "1.5"
 once_cell = "1.8"
 bitvec = "1"
+hex = "0.4"
+base64 = "0.13.0"
+tempfile = "3.3.0"
+prost-types = "0.9"
+tendermint-proto = "0.24.0-pre.1"
+prost = "0.9"
+ibc = { git = "https://github.com/penumbra-zone/ibc-rs.git", branch = "with-tendermintrs-24" }
+ibc-proto = { git = "https://github.com/penumbra-zone/ibc-rs.git", branch = "with-tendermintrs-24" }
+tendermint-light-client-verifier = "0.24.0-pre.1"
 
 [dev-dependencies]
 ed25519-consensus = "2"
 rand_core = "0.6"
+penumbra-tct = { path = "../tct" }
 
 [build-dependencies]
 vergen = "5"
