+++ component/src/ibc/component/channel.rs
@@ -1,6 +1,6 @@
-use crate::component::client::View as _;
-use crate::component::connection::View as _;
-use crate::event;
+use crate::ibc::component::client::View as _;
+use crate::ibc::component::connection::View as _;
+use crate::ibc::event;
 use anyhow::Result;
 use async_trait::async_trait;
 use ibc::core::ics02_client::client_consensus::AnyConsensusState;
@@ -25,7 +25,7 @@ use ibc::core::ics04_channel::packet::Packet;
 use ibc::core::ics24_host::identifier::ChannelId;
 use ibc::core::ics24_host::identifier::PortId;
 use penumbra_chain::genesis;
-use penumbra_component::{Component, Context};
+use crate::{Component, Context};
 use penumbra_proto::ibc::ibc_action::Action::{
     Acknowledgement, ChannelCloseConfirm, ChannelCloseInit, ChannelOpenAck, ChannelOpenConfirm,
     ChannelOpenInit, ChannelOpenTry, RecvPacket, Timeout,
