--- ibc/Cargo.toml
@@ -1,40 +1,0 @@
-[package]
-name = "penumbra-ibc"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-# Workspace dependencies
-penumbra-crypto = { path = "../crypto" }
-penumbra-chain = { path = "../chain" }
-penumbra-component = { path = "../component" }
-penumbra-storage = { path = "../storage" }
-penumbra-transaction = { path = "../transaction" }
-penumbra-proto = { path = "../proto" }
-tendermint = "0.24.0-pre.1"
-jmt = { git = "https://github.com/penumbra-zone/jellyfish-merkle.git", branch = "main" }
-
-# External dependencies
-hex = "0.4"
-base64 = "0.13.0"
-tokio = { version = "1.16", features = ["full"]}
-tempfile = "3.3.0"
-tracing = "0.1"
-prost-types = "0.9"
-tendermint-proto = "0.24.0-pre.1"
-once_cell = "1"
-serde_json = "1"
-serde = { version = "1", features = ["derive"] }
-anyhow = "1"
-prost = "0.9"
-async-trait = "0.1.52"
-ibc = { git = "https://github.com/penumbra-zone/ibc-rs.git", branch = "with-tendermintrs-24" }
-ibc-proto = { git = "https://github.com/penumbra-zone/ibc-rs.git", branch = "with-tendermintrs-24" }
-tendermint-light-client-verifier = "0.24.0-pre.1"
-metrics = "0.18"
-sha2 = "0.9"
-
-[dev-dependencies]
-penumbra-tct = { path = "../tct" }
